--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 **/.stack-work
 .stack/
+tensorflow-mnist-input-data/data/*.gz