# Before

```
benchmarking feedFetch/4 byte
time                 56.07 μs   (54.64 μs .. 57.31 μs)
                     0.995 R²   (0.993 R² .. 0.997 R²)
mean                 57.66 μs   (56.57 μs .. 58.74 μs)
std dev              3.851 μs   (3.350 μs .. 4.642 μs)
variance introduced by outliers: 68% (severely inflated)

benchmarking feedFetch/4 KiB
time                 228.8 μs   (227.4 μs .. 230.3 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 229.1 μs   (228.5 μs .. 230.5 μs)
std dev              3.016 μs   (1.902 μs .. 5.111 μs)

benchmarking feedFetch/4 MiB
time                 596.5 ms   (457.2 ms .. 661.2 ms)
                     0.994 R²   (0.984 R² .. 1.000 R²)
mean                 569.8 ms   (536.0 ms .. 589.7 ms)
std dev              30.62 ms   (0.0 s .. 34.45 ms)
variance introduced by outliers: 19% (moderately inflated)
```


# After

```
benchmarking feedFetch/4 byte
time                 58.26 μs   (57.03 μs .. 59.75 μs)
                     0.996 R²   (0.993 R² .. 0.998 R²)
mean                 61.08 μs   (59.17 μs .. 65.55 μs)
std dev              9.397 μs   (4.998 μs .. 17.20 μs)
variance introduced by outliers: 93% (severely inflated)

benchmarking feedFetch/4 KiB
time                 96.54 μs   (96.08 μs .. 97.07 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 97.85 μs   (97.21 μs .. 98.96 μs)
std dev              2.595 μs   (1.807 μs .. 4.076 μs)
variance introduced by outliers: 23% (moderately inflated)

benchmarking feedFetch/4 MiB
time                 126.0 ms   (125.4 ms .. 127.3 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 127.5 ms   (127.1 ms .. 128.0 ms)
std dev              720.2 μs   (580.4 μs .. 882.0 μs)
variance introduced by outliers: 11% (moderately inflated)
```