@fkm3 2016-11-13
FYI in case someone has peeked at this PR already: GHC was optimizing away most of the benchmark. I've fixed it and updated the numbers in the description.