@fkm3 2016-11-13
I've got a more involved change that makes a bigger difference here: https://github.com/tensorflow/haskell/commit/f176806df63c71eeee486e7ee9f7c640257f89bf